[PaTH] Prevent int32 index overflow for long sequences (#769)

+++ fla/ops/path_attn/parallel_path_bwd_inter_dkv.py
@@ -103,7 +103,7 @@ def parallel_path_bwd_dkv_kernel(
         b_delta = tl.load(p_delta, boundary_check=(0, ))
         b_l = tl.load(p_l, boundary_check=(0, ))
 
-        p_q = tl.make_block_ptr(q + ((bos * NUM_BLOCKS + idx_j) * HQ + i_hq) * K, (T, K),
+        p_q = tl.make_block_ptr(q + ((bos.to(tl.int64) * NUM_BLOCKS + idx_j) * HQ + i_hq) * K, (T, K),
                                 (HQ*K*NUM_BLOCKS, 1), (offset, 0), (BS, BK), (1, 0))
         b_q = tl.load(p_q, boundary_check=(0, 1))
         b_A = tl.dot(b_k, tl.trans(b_q).to(b_k.dtype))
